v4l2_backend: add query control function
Add function to perform a v4l2 query control ioctl
call using the control id received as an argument,
and return it through the 'value' output argument.

The VIDIOC_QUERYCTRL command queries a given control
set in the id field of a struct v4l2_queryctrl.
Driver may then return EINVAL if a control in
this range is not supported.

Signed-off-by: Albert Esteve <aesteve@redhat.com>